Compute detector error model of non-deterministic detectors (#22)
Closes #21 

Adds `get_detector_error_model` that converts observables into
detectors, uses stim to compute an error model, and convert detectors
back into observables.

There is one caveat. Stim will automatically insert `error(0.5) D3` for
non-deterministic detectors. This means that if the circuit had an error
mechanism that flips this observable `D3` and nothing else, the
probability for that error mechanism will be lost.

Generally, this should be acceptable because one would design circuits
to not have such error mechanisms (circuit distance 1). But we may in
the future consider computing the DEM ourselves or at least adding some
check and throwing an error for this case.

Sometimes stim may even remove gauges, e.g., in the case
```
  """
  R 0
  H 0
  X_ERROR(0.01) 0
  M 0
  OBSERVABLE_INCLUDE(0) rec[-1]
  """
```
Here, the DEM is simply empty, even though it should be `error(0.01)
L0`. If logical observables are removed by stim, the function in this PR
raises an error.